update context logic

+++ SessionManagement_service/internal/service/service.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 
+	"github.com/niktin06sash/MicroserviceProject/SessionManagement_service/internal/erro"
 	"github.com/niktin06sash/MicroserviceProject/SessionManagement_service/internal/logger"
 	"github.com/niktin06sash/MicroserviceProject/SessionManagement_service/internal/repository"
 	pb "github.com/niktin06sash/MicroserviceProject/SessionManagement_service/proto"
@@ -23,25 +24,52 @@ func NewService(repos *repository.Repository, log *logger.SessionLogger) *Servic
 }
 
 func (s *Service) CreateSession(ctx context.Context, req *pb.CreateSessionRequest) (*pb.CreateSessionResponse, error) {
+	requestID, ok := ctx.Value("requestID").(string)
+	if !ok || requestID == "" {
+		s.logger.Error("Request ID not found in context", zap.Error(erro.ErrorMissingRequestID))
+		return nil, erro.ErrorMissingRequestID
+	}
 	resp, err := s.sessionService.CreateSession(ctx, req)
 	if err != nil {
-		s.logger.Error("CreateSession Error", zap.Error(err))
+		s.logger.Error("ValidateSession Error",
+			zap.String("requestID", requestID),
+			zap.Error(err),
+		)
+		return nil, err
 	}
 	return resp, nil
 }
 
 func (s *Service) ValidateSession(ctx context.Context, req *pb.ValidateSessionRequest) (*pb.ValidateSessionResponse, error) {
+	requestID, ok := ctx.Value("requestID").(string)
+	if !ok || requestID == "" {
+		s.logger.Error("Request ID not found in context", zap.Error(erro.ErrorMissingRequestID))
+		return nil, erro.ErrorMissingRequestID
+	}
 	resp, err := s.sessionService.ValidateSession(ctx, req)
 	if err != nil {
-		s.logger.Error("ValidateSession Error", zap.Error(err))
+		s.logger.Error("ValidateSession Error",
+			zap.String("requestID", requestID),
+			zap.Error(err),
+		)
+		return nil, err
 	}
 	return resp, nil
 }
 
 func (s *Service) DeleteSession(ctx context.Context, req *pb.DeleteSessionRequest) (*pb.DeleteSessionResponse, error) {
+	requestID, ok := ctx.Value("requestID").(string)
+	if !ok || requestID == "" {
+		s.logger.Error("Request ID not found in context", zap.Error(erro.ErrorMissingRequestID))
+		return nil, erro.ErrorMissingRequestID
+	}
 	resp, err := s.sessionService.DeleteSession(ctx, req)
 	if err != nil {
-		s.logger.Error("DeleteSession Error", zap.Error(err))
+		s.logger.Error("ValidateSession Error",
+			zap.String("requestID", requestID),
+			zap.Error(err),
+		)
+		return nil, err
 	}
 	return resp, nil
 }
